- Fix unsigned->double, double->unsigned convs - Support Syncs in spillslot coalescer
[r15499]

+++ ir/be/test/fehler65.c
@@ -2,7 +2,7 @@ int randn = -1271796327;
 double value = 4294967295;
 
 int main(void) {
-	float res = (double) (randn % (unsigned int) value);
+	double res = (double) (randn % (unsigned int) value);
 	printf("Res: %f\n", res);
 	return 0;
 }
